[MODIFY] 파일 분할
