+++ src/modules/action/admin/interface/index.ts
@@ -8,3 +8,6 @@ export const QUESTION_LIST = 'ADMIN/QUESTION_LIST' as const;
 export const QUESTION_LIST_SUCCESS = 'ADMIN/QUESTION_LIST_SUCCESS' as const;
 export const QUESTION_LIST_FAILURE = 'ADMIN/QUESTION_LIST_FAILURE' as const;
 export const PAGE = 'ADMIN/PAGE' as const;
+export const QUESTION_LIST_DETAIL = 'ADMIN/QUESTION_LIST_DETAIL' as const;
+export const QUESTION_LIST_DETAIL_SUCCESS = 'ADMIN/QUESTION_LIST_DETAIL_SUCCESS' as const;
+export const QUESTION_LIST_DETAIL_FAILURE = 'ADMIN/QUESTION_LIST_DETAIL_FAILURE' as const;
