--- src/modules/middleware/admin/index.ts
@@ -1,5 +1,0 @@
-import adminList from './list';
-
-export default function* adminListSaga() {
-  yield [adminList()];
-}
