+++ src/modules/middleware/index.ts
@@ -1,7 +1,9 @@
 import { all } from 'redux-saga/effects';
 import post from './post';
-import adminListSaga from './admin/list';
+import adminQuestionSaga from './admin/list/Question';
+import adminReportPostSaga from './admin/list/ReportPost';
+import adminReportUserSaga from './admin/list/ReportUser';
 
 export default function* rootSaga() {
-  yield all([post(), adminListSaga()]);
+  yield all([post(), adminQuestionSaga(), adminReportPostSaga(), adminReportUserSaga()]);
 }
