+++ src/modules/reducer/admin/interface/list/Question.ts
@@ -7,4 +7,5 @@ export interface AdminQuestionListState {
   isHaveNextPage: boolean;
   isSuccessGetList: boolean | undefined;
   error: error;
+  description: string;
 }
