+++ src/modules/reducer/admin/list/QuestionState.ts
@@ -17,6 +17,7 @@ const initState: AdminQuestionListState = {
     message: '',
     type: '',
   },
+  description: '',
 };
 
 const adminQuestionListReducer = (
@@ -36,6 +37,7 @@ const adminQuestionListReducer = (
           list: action.payload,
           isSuccessGetList: true,
           isHaveNextPage: true,
+          description: ''
         };
       else
         return {
