+++ src/util/api/admin/list.ts
@@ -4,9 +4,7 @@ import { getRequestWithAccessToken } from '../default';
 export const getAdminQuestionList = async (accessToken: string, page: number) => {
   try {
     const request = getRequestWithAccessToken(accessToken);
-    const response = await request.get(`${uri.questionList}?page=${page}&size=5&sort=desc`);
-    console.log(response);
-
+    const response = await request.get(`${uri.questionList}?page=${page}&size=15&sort=desc`);
     return response;
   } catch (error) {
     throw error;
@@ -16,7 +14,7 @@ export const getAdminQuestionList = async (accessToken: string, page: number) =>
 export const getAdminReportUserList = async (accessToken: string, page: number) => {
   try {
     const request = getRequestWithAccessToken(accessToken);
-    const response = await request.get(`${uri.reportUserList}?page=${page}&size=5&sort=desc`);
+    const response = await request.get(`${uri.reportUserList}?page=${page}&size=15&sort=desc`);
     return response;
   } catch (error) {
     throw error;
@@ -26,7 +24,7 @@ export const getAdminReportUserList = async (accessToken: string, page: number) 
 export const getAdminReportPostList = async (accessToken: string, page: number) => {
   try {
     const request = getRequestWithAccessToken(accessToken);
-    const response = await request.get(`${uri.reportPostList}?page=${page}&size=5&sort=desc`);
+    const response = await request.get(`${uri.reportPostList}?page=${page}&size=15&sort=desc`);
     return response;
   } catch (error) {
     throw error;
